chore(pyamber): move tests to src/test/python to match Maven layout
Aligns Python with the Scala source-set layout already used by amber:
production code in `src/main/python`, tests in `src/test/python`. The
forward-looking comment in `amber/build.sbt` already anticipated this
move.

- `git mv` 48 `test_*.py` files preserving package structure.
- Convert 7 test files from relative imports (`from .x import Y`) to
  absolute paths now that the test tree is no longer a sibling of the
  production module.
- Drop the hardcoded `core.util.expression_evaluator.test_expression_evaluator`
  prefix in test_evaluate_object_expression — use `A.__module__` so the
  assertion tracks whatever the test file's import path resolves to.
- Move `pyproject.toml` from `src/main/python/` to `amber/` so a single
  config sits at the module root (mirroring `amber/build.sbt` for Scala).
- Add `[tool.pytest.ini_options]` with `pythonpath = ["src/main/python"]`,
  `testpaths = ["src/test/python"]`, and `--import-mode=importlib` so
  test modules don't need `__init__.py` mirrors and don't collide with
  the production package namespace.
- CI: invoke ruff + pytest from `amber/` with explicit `src/{main,test}/python`
  paths; codecov uploads from `./amber/coverage.xml`.
- `bin/fix-format.sh` collapses to a single `cd amber && ruff format
  src/main/python src/test/python`.

Closes #4943
