+++ amber/build.sbt
@@ -56,10 +56,10 @@ Compile / unmanagedSourceDirectories += baseDirectory.value / "src" / "main" / "
 
 // `amber/src/test/integration` holds Scala specs that exercise both
 // Scala and Python end-to-end (tagged @org.apache.texera.amber.tags.IntegrationTest).
-// Sits next to `src/test/scala` and `src/test/java`; a future `src/test/python`
-// can drop in the same way. Adding it to Test/unmanagedSourceDirectories means
-// scalafmtCheckAll / scalafixAll --check naturally cover these sources, and
-// the AMBER_TEST_FILTER env var below routes which tagged subset runs.
+// Sits next to `src/test/scala`, `src/test/java`, and `src/test/python`.
+// Adding it to Test/unmanagedSourceDirectories means scalafmtCheckAll /
+// scalafixAll --check naturally cover these sources, and the
+// AMBER_TEST_FILTER env var below routes which tagged subset runs.
 Test / unmanagedSourceDirectories += baseDirectory.value / "src" / "test" / "integration"
 
 // Test-filter switch driven by the AMBER_TEST_FILTER env var so the
