+++ amber/src/test/python/core/architecture/managers/test_console_message_manager.py
