+++ amber/src/test/python/core/architecture/managers/test_tuple_processing_manager.py
