+++ amber/src/test/python/core/models/test_operator.py
