+++ amber/src/test/python/core/util/console_message/test_replace_print.py
