+++ amber/src/test/python/core/util/thread/test_atomic.py
